Updated Graphs Implementation & Renamed Graphs Files
Node Id's now indicate where they are located in an array, so they function both as an nodeId and an index.

If the node is invalidated, it's not removed, but rather the nodeId is set to InvalidNodeID.

Whenever a new node is created, it will search for an Node with an invalidNodeId and replaces that node.

Removed trivial getters/setters and made the variables public instead.

Remove Graph from GraphXXX files, since it's already in a graphs folder

+++ Graphs/HeuristicFunctions.hpp
@@ -15,6 +15,7 @@ namespace bae::Graphs::HeuristicFunctions
 {
     #pragma warning(push)
     #pragma warning(disable : 4505)
+
     // Manhattan distance
     static float Manhattan(float x, float y)
     {
@@ -24,13 +25,13 @@ namespace bae::Graphs::HeuristicFunctions
     // Euclidean distance
     static float Euclidean(float x, float y)
     {
-        return static_cast<float>(std::sqrtf(x * x + y * y));
+        return std::sqrtf(x * x + y * y);
     }
 
     // Sqrt Euclidean distance
     static float SqEuclidean(float x, float y)
     {
-        return static_cast<float>(x * x + y * y);
+        return x * x + y * y;
     }
 
     // Octile distance
